% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-simem.R
\name{get_simem_variables}
\alias{get_simem_variables}
\title{Get SIMEM Variable Collection}
\usage{
get_simem_variables()
}
\value{
Tibble with available variables and all their metadata fields.
List columns are preserved for fields like dimensions.
}
\description{
Get list of available SIMEM variables with all metadata
}
