% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-simem.R
\name{SIMEMCatalog}
\alias{SIMEMCatalog}
\title{SIMEM Catalog Client}
\description{
R6 class for accessing SIMEM catalog information
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{catalog_type}}{Type of catalog (datasets or variables)}

\item{\code{data}}{Catalog data}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SIMEMCatalog-new}{\code{SIMEMCatalog$new()}}
\item \href{#method-SIMEMCatalog-get_data}{\code{SIMEMCatalog$get_data()}}
\item \href{#method-SIMEMCatalog-clone}{\code{SIMEMCatalog$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SIMEMCatalog-new"></a>}}
\if{latex}{\out{\hypertarget{method-SIMEMCatalog-new}{}}}
\subsection{Method \code{new()}}{
Initialize the SIMEM catalog client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SIMEMCatalog$new(catalog_type)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{catalog_type}}{Type of catalog: "datasets" or "variables"}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new SIMEMCatalog object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SIMEMCatalog-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-SIMEMCatalog-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get catalog data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SIMEMCatalog$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with catalog information
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SIMEMCatalog-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SIMEMCatalog-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SIMEMCatalog$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
