% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/client-simem.R
\name{SIMEMVariable}
\alias{SIMEMVariable}
\title{SIMEM Variable Client}
\description{
R6 class for accessing specific SIMEM variables
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{cod_variable}}{Variable code}

\item{\code{start_date}}{Start date for data retrieval}

\item{\code{end_date}}{End date for data retrieval}

\item{\code{version}}{Version of the variable}

\item{\code{variable_info}}{Variable configuration information}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-SIMEMVariable-new}{\code{SIMEMVariable$new()}}
\item \href{#method-SIMEMVariable-get_data}{\code{SIMEMVariable$get_data()}}
\item \href{#method-SIMEMVariable-get_stats}{\code{SIMEMVariable$get_stats()}}
\item \href{#method-SIMEMVariable-clone}{\code{SIMEMVariable$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SIMEMVariable-new"></a>}}
\if{latex}{\out{\hypertarget{method-SIMEMVariable-new}{}}}
\subsection{Method \code{new()}}{
Initialize the SIMEM variable client
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SIMEMVariable$new(cod_variable, start_date, end_date, version = 0)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{cod_variable}}{Character string with variable code}

\item{\code{start_date}}{Start date in YYYY-MM-DD format}

\item{\code{end_date}}{End date in YYYY-MM-DD format}

\item{\code{version}}{Version of the variable (default: 0)}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new SIMEMVariable object
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SIMEMVariable-get_data"></a>}}
\if{latex}{\out{\hypertarget{method-SIMEMVariable-get_data}{}}}
\subsection{Method \code{get_data()}}{
Get variable data
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SIMEMVariable$get_data()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Data frame with variable data
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SIMEMVariable-get_stats"></a>}}
\if{latex}{\out{\hypertarget{method-SIMEMVariable-get_stats}{}}}
\subsection{Method \code{get_stats()}}{
Get variable statistics
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SIMEMVariable$get_stats()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
List with variable statistics
Read JSON configuration for variables
Process variable data based on configuration
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-SIMEMVariable-clone"></a>}}
\if{latex}{\out{\hypertarget{method-SIMEMVariable-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{SIMEMVariable$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
